FCBH-1175 Temp playlist support
Added audio filesets to the payload of both the highlights and the bible text search results with minimal additions to querying.  We are not assured to have audio but we are following this up with an endpoint to check if audio is avaialble for a specific verse level location in the bible.