FCBH-1337 Fix language ordering
Languages were only being ordered by their name in the languages table.  This did not acuratly reflect how these languages were eventually presented with their name either being that value or the value of the name in language_translations.  Therefore the order was flawed.  Using a raw sort and an ifnull statement, the same procedure for the transformer was applied to the ordering.